# src/swagger.json
{
  "basePath": "/",
  "consumes": ["application/json"],
  "info": {
    "title": "Passphrase Launchpad Server",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    }
  },
...
  "definitions": {
...
    "Companies": {
      "properties": {
        "name": {
          "type": "string",
          "default": null,
          "required": true
        },
        "description": {
          "type": "string",
          "default": null,
          "required": true
        },
        "createdAt": {
          "type": "number",
          "default": null
        },
        "category": {
          "type": "string",
          "default": null,
          "required": true
        },
        "websiteUrl": {
          "type": "string",
          "default": null
        },
        "iconUrl": {
          "type": "string",
          "default": null,
          "required": true
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "detail": {
                "type": "string",
                "default": null
              },
              "description": {
                "type": "string",
                "default": null
              }
            }
          }
        },
        "fundingTeam": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "memberImg": {
                "type": "string",
                "default": null,
                "required": true
              },
              "name": {
                "type": "string",
                "default": null,
                "required": true
              },
              "surname": {
                "type": "string",
                "default": null
              },
              "position": {
                "type": "string",
                "default": null,
                "required": true
              },
              "linkedinLink": {
                "type": "string",
                "default": null
              }
            }
          }
        },
        "socialMedia": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "media": {
                "type": "string",
                "default": "telegram",
                "required": true
              },
              "url": {
                "type": "string",
                "default": null,
                "required": true
              }
            }
          }
        }
      },
      "required": ["name", "description", "createdAt", "category", "iconUrl"],
      "type": "object"
    },
    "FundingPools": {
      "properties": {
        "slug": {
          "type": "string",
          "unique": true,
          "dropDups": true,
          "required": true,
          "lowercase": true,
          "trim": true
        },
        "description": {
          "type": "string",
          "default": null,
          "required": true
        },
        "auctionStart": {
          "type": "number",
          "default": null
        },
        "auctionEnd": {
          "type": "number",
          "default": null,
          "required": true
        },
        "pricePerToken": {
          "type": "number",
          "default": null,
          "required": true
        },
        "vesting": {
          "type": "string",
          "default": null,
          "required": true
        },
        "launchpads": {
          "type": "string",
          "default": null,
          "required": true
        },
        "backers": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {}
          },
          "default": null
        },
        "minAmount": {
          "type": "number",
          "default": null,
          "required": true
        },
        "maxAmount": {
          "type": "number",
          "default": null,
          "required": true
        },
        "capacity": {
          "type": "number",
          "default": null,
          "required": true
        },
        "saleType": {
          "type": "string",
          "default": null,
          "required": true
        },
        "status": {
          "type": "string",
          "default": null
        },
        "company": {
          "type": "string",
          "ref": "Company",
          "required": true,
          "default": null
        },
        "templateId": {
          "type": "string",
          "default": null
        },
        "saftFiles": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "saftId": {
                "type": "string",
                "default": null
              },
              "ownerId": {
                "type": "string",
                "default": null,
                "required": true,
                "ref": "User"
              },
              "isValid": {
                "type": "boolean",
                "default": true
              },
              "procedureId": {
                "type": "string",
                "default": null
              },
              "signatureId": {
                "type": "string",
                "default": null
              },
              "createdAt": {
                "type": "number",
                "default": null
              }
            }
          },
          "default": null
        },
        "walletAddress": {
          "type": "string",
          "default": null
        },
        "contractAddress": {
          "type": "string",
          "default": null
        },
        "title": {
          "type": "string",
          "required": true,
          "default": null
        },
        "referrerFee": {
          "type": "string",
          "required": true,
          "default": null
        },
        "createdAt": {
          "type": "number",
          "required": true,
          "default": null
        }
      },
      "required": [
        "slug",
        "description",
        "auctionStart",
        "auctionEnd",
        "pricePerToken",
        "vesting",
        "launchpads",
        "minAmount",
        "maxAmount",
        "capacity",
        "company",
        "title",
        "referrerFee",
        "createdAt"
      ],
      "type": "object"
    },
...
  },
  "paths": {
...
    "/v1/companies/search": {
      "get": {
        "tags": ["Company"],
        "operationId": "searchCompanies",
        "summary": "Search Company",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Companies"
                  }
                },
                "count": {
                  "type": "number",
                  "default": 0
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "searchCompaines",
              "type": "object",
              "required": ["success"]
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 0
          }
        ]
      }
    },
...
    "/v2/companies/search": {
      "get": {
        "tags": ["Company"],
        "operationId": "searchCompanies",
        "summary": "Search Company",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Companies"
                  }
                },
                "count": {
                  "type": "number",
                  "default": 0
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "searchCompaines",
              "type": "object",
              "required": ["success"]
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 0
          }
        ]
      }
    },
...
    "/v1/companies": {
      "post": {
        "tags": ["Company"],
        "operationId": "registerCompany",
        "summary": "Register Company",
        "description": "Endpoint for register company",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful registered",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "registerCompany",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "tags": ["Company"],
        "operationId": "getAllCompanies",
        "summary": "Get All Companies",
        "description": "Endpoint for get all companies",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "registerCompany",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": []
      }
    },
    "/v2/companies": {
      "post": {
        "tags": ["Company"],
        "operationId": "registerCompany",
        "summary": "Register Company",
        "description": "Endpoint for register company",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful registered",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "registerCompany",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "tags": ["Company"],
        "operationId": "getAllCompanies",
        "summary": "Get All Companies",
        "description": "Endpoint for get all companies",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "registerCompany",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": []
      }
    },
    "/v1/companies/{id}": {
      "get": {
        "tags": ["Company"],
        "operationId": "getCompanyById",
        "summary": "Get Company By Id",
        "description": "Endpoint for get company by Id",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "getCompanyById",
              "type": "object"
            }
          },
          "401": {
            "description": "Company does not exist!"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Company"],
        "operationId": "updateCompany",
        "summary": "Update Company",
        "description": "Endpoint for update company",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully updated",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "updateCompany",
              "type": "object"
            }
          },
          "401": {
            "description": "Company does not exist!"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Company"],
        "operationId": "deleteCompany",
        "summary": "Delete Company",
        "description": "Endpoint for delete company",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "boolean",
                  "default": true
                }
              },
              "title": "deleteUser",
              "type": "object",
              "required": []
            }
          },
          "401": {
            "description": "Company does not exist!"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/companies/{id}": {
      "get": {
        "tags": ["Company"],
        "operationId": "getCompanyById",
        "summary": "Get Company By Id",
        "description": "Endpoint for get company by Id",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "getCompanyById",
              "type": "object"
            }
          },
          "401": {
            "description": "Company does not exist!"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Company"],
        "operationId": "updateCompany",
        "summary": "Update Company",
        "description": "Endpoint for update company",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully updated",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "updateCompany",
              "type": "object"
            }
          },
          "401": {
            "description": "Company does not exist!"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Companies",
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Company"],
        "operationId": "deleteCompany",
        "summary": "Delete Company",
        "description": "Endpoint for delete company",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "boolean",
                  "default": true
                }
              },
              "title": "deleteUser",
              "type": "object",
              "required": []
            }
          },
          "401": {
            "description": "Company does not exist!"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/fundingPools/company/{companyId}": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getCompanyFundingPools",
        "summary": "Get Company FundingPools",
        "description": "Endpoint for get company fundingPools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "companyId",
            "schema": {
              "properties": {
                "companyId": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["companyId"]
            }
          }
        ]
      }
    },
    "/v2/fundingPools/company/{companyId}": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getCompanyFundingPools",
        "summary": "Get Company FundingPools",
        "description": "Endpoint for get company fundingPools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "companyId",
            "schema": {
              "properties": {
                "companyId": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["companyId"]
            }
          }
        ]
      }
    },
...
  },
  "produces": ["application/json"],
  "swagger": "2.0",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "body",
      "description": "This API requires authentication using a bearer token. Please include the token in the 'Authorization' header of your requests. ex: Bearer [space] <your-token>"
    }
  }
}